requirements.txt
# ==========================================
# KATITA-POS - Backend Dependencies
# Python 3.12.3
# ==========================================

# Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Security
bcrypt==4.1.2
python-dotenv==1.0.0

# Validation & Utilities
marshmallow==3.20.1
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
coverage==7.3.2

# Development
black==23.12.1
flake8==6.1.0
autopep8==2.0.4

# Production Server
gunicorn==21.2.0
